fixing tenant cleanup issue with Ghost deployer and WSAS-931
git-svn-id: https://svn.wso2.org/repos/wso2/carbon/platform/trunk@135824 a5903396-d722-0410-b921-86c7d4935375

+++ components/webapp-mgt/org.wso2.carbon.webapp.mgt/src/main/java/org/wso2/carbon/webapp/mgt/WebApplication.java
@@ -277,20 +277,28 @@ public class WebApplication {
                     getUrlMappingsPerApplication(this.context.getName());
             Engine engine = DataHolder.getCarbonTomcatService().getTomcat().getEngine();
             Context hostContext;
+            Host host;
             for (String hostName : mappings) {
-                hostContext = (Context) engine.findChild(hostName).findChild("/");
-                if (nameOfOperation.equalsIgnoreCase("start")) {
-                    start(hostContext);
-                } else if (nameOfOperation.equalsIgnoreCase("stop")) {
-                    stop(hostContext);
-                } else if (nameOfOperation.equalsIgnoreCase("reload")) {
-                    reload(hostContext);
-                } else if (nameOfOperation.equalsIgnoreCase("lazyunload")) {
-                    lazyUnload(hostContext);
-                    DataHolder.getHotUpdateService().removeHost(hostName);
-                } else if (nameOfOperation.equalsIgnoreCase("delete")) {
-                    DataHolder.getHotUpdateService().deleteHost(hostName);
+                host = (Host) engine.findChild(hostName);
+                if(host != null) {
+                    hostContext = (Context)host.findChild("/");
+                    if(hostContext != null) {
+                        if (nameOfOperation.equalsIgnoreCase("start")) {
+                            start(hostContext);
+                        } else if (nameOfOperation.equalsIgnoreCase("stop")) {
+                            stop(hostContext);
+                        } else if (nameOfOperation.equalsIgnoreCase("reload")) {
+                            reload(hostContext);
+                        } else if (nameOfOperation.equalsIgnoreCase("lazyunload")) {
+                            lazyUnload(hostContext);
+                            DataHolder.getHotUpdateService().removeHost(hostName);
+                        } else if (nameOfOperation.equalsIgnoreCase("delete")) {
+                            DataHolder.getHotUpdateService().deleteHost(hostName);
+                        }
+                    }
+
                 }
+
             }
 
         }
